+++ Source/CSBackend/Rendering/OpenGL/Model/GLMesh.cpp
@@ -89,7 +89,7 @@ namespace CSBackend
                 auto normalised = GLMeshUtils::IsNormalised(elementType);
                 auto offset = reinterpret_cast<const GLvoid*>(u64(vertexFormat.GetElementOffset(i)));
                 
-                glShader->SetAttribute((int)elementType, numComponents, type, normalised, vertexFormat.GetSize(), offset);
+                glShader->SetAttribute((u8)elementType, numComponents, type, normalised, vertexFormat.GetSize(), offset);
             }
             
             for (u32 i = vertexFormat.GetNumElements(); i < m_maxVertexAttributes; ++i)
